requirements.txt
Django==1.10.2
django-appconf==1.0.2
django-imagekit==3.3
pilkit==1.1.13
Pillow==3.4.1
six==1.10.0
